Fix enumerate() and getPoint() of CGPath.
Fix enumerate(): Not call getPoint(0) to avoid re-entry problem.
In addition, for kCGPathElementCloseSubpath, a line should to be appended from the current point to the starting point of the current subpath instead of the first point of the whole path.

Fix getPoint() to avoid missing points due to the difference between element and point of bezier segments.